requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
deepface==0.0.93
scikit-learn==1.4.2
